created a rosservice for updating the tree
Modified the htt_viz NodeView.py so that the NodeView class also publishes a service which can be called by the HTT itself to change the color of existing nodes. For now the service only has an empty callback function so it doesn't do anything but the service is called 'update_htt' and we can use it like a black box until there is functionality in there. Also added an asynchronous spinner which ends up a member of the NodeView class, this is useful to make sure that messages can be published by this node and sent to this node.