Tree Class
Made a small update to the node class by adding a parent, and then also made a tree class which holds a dictionary of nodes and as well has them connected in a classical manner. Currently the tree class has an untested add and remove method so before this can be merged those will need to be fleshed out. Finally, the constructor should reasonably take in whatever we get out of the yaml. or perhaps we can always just make trees by using the addchild method in the translator class