Further testing done, and drawing fixed.
I think I covered all of the cases for the remove node function and tested both of them, then added a function to recursively draw the tree object from a root node, and a function to draw the whole tree. Then finally, I updated the NodeView object to use the tree in an example so I think everything works and I hope we don't have to revisit this class.