filled in the callback
Made some minor changes to how the Nodes are drawn so that they now draw with a particularly colored brush. Also changed the callback so that it updates the given node with the given activation_potential and the given color and refreshes. this can be tested with a simple command line cmd "rosservice call update_htt "root" true 0.0" while the htt_viz node is running.