Compiles!
Finally, got the main run file to compile on my laptop, so one of the things I'll do next is put the dependencies for compiling and running in the next READ_ME.md push, although for now it's ok if we can only compile on my machine for the demo. Created a launch file, a test tree, a place for both, and started a new DummyBehavior although it isn't included in the CMAKE yet. Also included all the messages necessary for the Tree to run, updated the CMAKELISTS and the package.xml and generally am moving towards writing the send_update algo in the node.cc file.