Update stuff done on the tree side.
Created a working tree, and now all of the nodes in the task tree will call the update service of htt_viz whenever their basic information has changed. It's and O(1) check in the tree code so it shouldn't slow things down, but it does it every cycle so I'm mildly worried about the efficiency of it. For now, however, everything works great and essentially the behaviors turn green whenever they are active. Also slightly changed test_0.yaml to include a single dummybehavior. In order to test things open the window in one console, and then run the launch file for the task tree in the other window, and the nodes should update. In the future we will be able to load task trees dynamically in the main window however for now I have it hard coded to match test_0.yaml.