Mostly done but Mostly Ugly
Added functionality to the menu button which starts the debug console. It now opens a window which only contains a text field which echoes the information that it receives from a certain node over the ros_out topic. We should be able to decide the color and the color of the font so we can even mimic the colors that are shown on linux but for now I'm taking fuction over form and pushing it.