Probably not helpful
This commit enables the main frame to handle key events, and for now all it does is print what keycode it received. I was planning on using this for the shortcuts however those are already working so that ends up being redundant. For now this branch will probably die but in case the KeyEvent code becomes useful I'll keep this copy around