Saving Done.
Implemented a function to dump the yaml version of a tree into an inputted file in the NodeView class. The intention of this is to have the Main frame simply pass in an opened file which then Node View dumps into, so far the created yaml files work well with the architecture (although they are dumped in alphabetical order which i wasn't sure would work). As well, the x y coordinates at the time of saving are saved in the yaml so loading works as expected and essentially loads whatever the user last had on the gui.