Some Functionality Done
The basics are there now, and the basis as well. The Stack works and is able to run functions with methods for other objects, and the functionality has been added for the add function so there are a few things left. Removing a node, and Moving nodes need undo functionality. Then as well, the star needs to go away if we empty the stack of changes. Saving the tree needs to empty the undo stack? or perhaps just needs to get rid of the star (not sure). And the redo stack has to somehow do the opposite of what the undo stack undid.

Also, each method in the tree is going to be changes (RemoveNode is already  done), so that they take in a list of arguments rather than a traditional list of parameters, and that way we can just store the arguments list in the stack. Also undo is added with functionality in the main file, although that should already be good to go.