I'm bad at stacks apparently
Fixed a couple bugs, removed some prints, and added a helpful method.

First, I was adding to the back of the stack instead of the front so that is fixed. Next, I stopped the undo from running if nothing is popped. I also changed the RemoveNodeRec function so that it conforms to the new args list I added to the RemoveNode. Finally, I commented up the Stack.py file so that it is more legible. Also, I'm slowly changing to use python conventions rather than cpp ones so some "== None"s have been changed to "is None" and so on. Also, changing my comments to use the conventions of a space after the #.