Some redo
Changed the StackNode class to ActionNode, and made it into a tuple of an undo and redo action. Then as well changed it so that a single Boolean switches it between which to use. Also, added functionality for the redo button by creating another instance of the stack such that when something is popped off of the undo stack it is popped onto the redo stack. Also changed some of the comments in the Stack.py file so that hopefully it reads a little better, changed some code for legibility, and as well changed some things for conventions of python.