Another Layer of abstraction
Added another class so that Action nodes only contain lists of function calls. This will allow them to simply run all of the function calls that make up a single action rather than a single function call per node in the stack. This is important for the undo of a remove node as removing a node removes all of its children and therefore adding them back would require as many nodes as there were children.