Adjusted RemoveNode Function
Ok so far as I know the remove Node function has now been adjusted so that it works with the undo-redo functionality. Essentially, a list is passed through all calls to remove node and is used to keep track of everything that is removed from the tree in terms of what functions need to be called to reverse and redo it. There is one thing currently unfinished which is that the children of a node will always be added in the reverse order list-wise although the tree will still look the same which could potentially cause problems with then nodes (the bug can be seen in what is saved to a yaml file). Don't know how I'm going to fix it for now, although I think I'm going to insert at a different position based on the knowledge of what child a node is in the list for its parent.