just little stuff
Changed it so that on opening a file it now clears the undo and redo stack. To that end added a clear function to the Stack class and also fixed some bugs where I hadn't updated the calls to Add and Remove Node. I also fixed the previously mentioned bug where the indices of the children would reverse, by adding a parameter in the args list which counts which child it is based on the index the for loop in RemoveNodeRec has. Also added comments for functions where parameters are unclear (specifically add and remove node in NodeView).