Movement can undo-redo now
Added a movenode function so that we can programmatically move them without the mouse. I think that the way they move is better for now so no need to change it elsewhere I'm only using the function in the undo to snap them back to where they were. That's pretty much all we need for now with the undo-redo stuff because we can't do very many things, however, there are some things we will need as we move further. When we can select groups of nodes we need to update the movement function so that we get a list of nodes that need to move and then the undo-redo stuff will work fine. And as well when we can edit nodes we'll have to create functionality for the undo-redo for that. Editting paramters as well etc.