Just comments on the Tree class
Just added a boatload of comments to the Tree class that hopefully allows it to be read by other people. Will move this to the main branch and then begin working on commenting the rest of the code before I continue adding functionality so that it is easier to add to for people other than me.