Added a functional test and removed a bug... and noticed another bug that I don't plan to fix
The test added is a python file which is not meant to run through ROS that simply makes a randomized tree that is approximately 951 nodes large and saves it to the disk. If it runs on the architecture then the save function is working. The bug I fixed was a bug in which the architecture would not create dummy behaviors and would instead create nothing and do nothing (simply took out an if statement that shouldn't have been there). Finally, noticed a bug that if the number of nodes is over 1000 then the hardcoded solutions for 3 digits fail. This is across the architecture and the htt_viz code itself and is pretty low danger because trees that large aren't really used. So I won't fix it for now but I have noted it.