added a copy constructor and a test for Tree
Further extending the functionality of the Tree class, I added a copy constructor in the form of the copy method in Tree which functions by utilizing the rec_copy method above it. The test itself makes a tree of some size, copies it to another object, and then checks if the two objects are equal (the test does not work unless the equality is working).