Added and Tested a function. And moved some code around
Added a findNodeByName and a recursive helper to the Tree class to facilitate the replacement of the node_dict behavior as we move forward. find_test.py is where the test for the function is and the function passed. Finally moved the Tree and Node class descriptions to Tree.py and removed them from NodeView so that they could be tested without preventing the program from using ros.