All tests for data structures passed, and node_dict removed
As stated undo_redo_test is now a pass with the advent of the better way of storing the tree. Equality is slightly changed so that the order of children does not matter, however, it needs to be preserved so this might need to be changed to do an insert instead of disregarding it in the future. Finally, all references to node-dict in the data structures are gone and so is node_dict therefore the next work on this branch will be dedicated to essentially fixing all of the rest of the code to not rely on node_dict. As well, we could do with changing where the tests are saved.