Fixed equality and preserved order with undo_redo instead
Equality was running in polynomial time and was going super slow as a result, so I changed it back to how we checked for equality before. This means that trees with nodes that are in different orders are no longer equal. To support this removing a node now saves the index that it was at in order to insert it back into the right spot upon undoing which preserves the order of nodes.