Changed some things and made sure it runs
Double checked and fixed all the bugs in the main.py and NodeView.py files that I created when I was changing how the Tree and Node classes worked during bugfixing, so the behavior is back to how it was before. Also, added a member to Node that contains the number of children to fix a bug where the tree would remove a subtree and count as only losing a single child. This should work, although it isn't that hard to fix if it isn't maintained correctly. Now fixing a problem where 2 nodes with the same name can exist in the Tree if you add a node-> remove another node -> add a node (the two added nodes will have the same node_num).