Bugfix
Fixed a bug where the numbers of the nodes added were not the next free number, and were able to overlap. Now the numbers should never overlap due to a membership array inside of the tree class saving the numbers which have been used already. New numbers are received from the getNextNum Function, and are maintained in the add and remove node functions with help from rec_num_maintainer, which takes in a ptr to a subtree and flips all of the bools in the array to the given value across the whole subtree recursively.