Merge Msg
So I merged in test-bugfix because it has all of the code that was changed on the backend. All of that code was already tested with the python files labeled test, and has passed those tests so it should be completely stable here. As well Tree, Node, and Stack.py should all be wx agnostic except that Node and tree both still have draw functions that will need to be changed.  Automated tests might be made in the future.