skeleton and basic read test
Just created a super basic skeleton of the Param class and inheritors and they will probably remain quite simple, however, they need to output to a yaml file as well and hold values. Also wrote the first test for the classes which will hold these params which is the read function and its corresponding test. Going to make sure that this can do a round trip next, move the test out of the library file, and as well do the write function.