Small changes
Essentially just set things up so that the ros service that HTT_VIZ uses will work again, although it can't work until the GUI gives me some sort of redraw function so that we can have blinky again. Looking through the Save function everything should be placed in the same order so far as I know when we save children, so please just make a test and we will fix if needed. In terms of the load function that probably is where the problem was in the demo but again I'd prefer to work on that after the GUI gets pushed essentially. Hopefully all ros things have been commented out for now although that is only because the backend is not properly hooked up, the idea is that we should just be able to uncomment them and then it should work.