ROSNode works again
Blinky with the behaviors works when they are running again although I made them look awful. Made 2 color change methods in the QGraphics node for this and called it a day so changes to the colors need to happen in that file. Also the old trees no longer load correctly which is my fault because I have since changed the name of the behaviors and it doesn't like that. Also changed some stuff around in the architecture itself so the nodetypes have the correct numbers and therefore load correctly.