Loading is done kinda
The loading function now loads the correct file paths for the node types required by each of the trees. And also loaded the parameters for each node correctly. Also changed the edit node window so it saves the parameters which ended up being like ~10 lines. Also changed some stuff on how things are saved in order to make loading make sense, so once again older yaml trees won't load. This brings up the last problem with loading in that the unhappy path where they don't load the correct file type or a previously saved tree it just won't load whereas it should essentially just port the old description to the new description (e.g. if no paths are given to behaviors then prompt the user to give all the files, if the parameters aren't saved then prompt the user and say default values are loaded, etc). Also nodes do not load with the previous coord values, should be an easy fix.