Silly Changes
So I was running this on linux and for some reason the enums weren't working so in htt_viz and QGraphicsTaskTreeNode.py I had to change the explicit enums to what their values actually are. So far as I can tell the program runs the same, however, so sincerely I'm sorry and we can get rid of it in merges. Other changes include the basic setup stages for how the draw behavior is going to work. I just wanted to push before I did too much work.