Basics are done
Basics of the draw behavior are set up. I quickly made a tree to test things with, and as well modified the loadtree file to load draw behaviors. Nothing talks to moveit yet but the messages are pretty much set up so that they just need to be sent over and it should be fine. Also thinking of another behavior to set things up so that the draw-behavior stuff can work, but essentially we'll see