Added the basic draw behavior
Basically now there's a working draw behavior that has been tested in sim and is added by default for the htt_viz program. It has a dependency that it needs the pr2_moveit_config library on the ROS_PATH and both launch files needs to be run for the behavior to work which is basically unavoidable and expected. Right now the behavior just spells my name because I have a god complex, and has been tested and is working in sim. Next step is to test it in the real world and hope for the best. Also grab behavior just moves to a pose coordinate (I used this for testing) so it definitely doesn't grab anything and I might remove it.